requirements.txt
fastapi
uvicorn
python-multipart
nemo_toolkit[asr]
